Separate test from AngleMarker class
